Checkbox moved to new component folder, Forms, and is now used by create training plan page
